Final Cucumber

+++ src/main/java/stepDefinitions/MasterHooks.java
@@ -4,6 +4,7 @@ import core.DriverFactory;
 import cucumber.api.Scenario;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
+import pageObjects.BasePage;
 
 public class MasterHooks extends DriverFactory {
 
@@ -11,11 +12,24 @@ public class MasterHooks extends DriverFactory {
 	public void setup() {
 		driver = getDriver();
 	}
-
+	
 	@After
 	public void tearDownAndScreenshotOnFailure(Scenario scenario) {
-		driver.manage().deleteAllCookies();
-		driver.quit();
-		driver = null;
+		try {
+			if (driver != null && scenario.isFailed()) {
+				
+				BasePage.captureScreenshot();
+				driver.manage().deleteAllCookies();
+				driver.quit();
+				driver = null;
+			}
+			if (driver != null) {
+				driver.manage().deleteAllCookies();
+				driver.quit();
+				driver = null;
+			}
+		} catch (Exception e) {
+			System.out.println("Methods failed: tearDownAndScreenshotOnFailure, Exception: " + e.getMessage());
+		}
 	}
 }
